[BE] refactor: 생성자 public -> private으로 변경, domain redis 구분 제거 (#10)
